WIP. Not working. Committing solely as a record.
I'm not convinced i even want to do this. From a blog post
(https://www.liwen.id.au/photoswipe/), it says photoswipe loads
all the images when you open the gallery in the lightbox, which is
not what i want. I want to load an image at a time.

Filing under "probably not worth it"!